ai-layer/requirements.txt
boto3>=1.28.0
pydantic>=2.0.0
python-dotenv>=1.0.0
PyJWT>=2.8.0
